+++ web/src/app/[locale]/(main)/[camera]/playback/page.tsx
@@ -1,12 +1,8 @@
 'use client'
 
 import { components } from '@/api/schema'
-import CameraPlayer, {
-  CameraPlayerHandle,
-} from '@/components/camera/camera-player'
-import CameraTimeline, {
-  CameraTimelineHandle,
-} from '@/components/camera/camera-timeline'
+import CameraPlayer, { CameraPlayerHandle } from '@/components/camera/camera-player'
+import CameraTimeline, { CameraTimelineHandle } from '@/components/camera/camera-timeline'
 import { Button } from '@/components/ui/button'
 import { Slider } from '@/components/ui/slider'
 import { openapi } from '@/lib/http'
@@ -20,7 +16,7 @@ import {
   CirclePlay,
   Fullscreen,
   Volume2,
-  VolumeOff,
+  VolumeOff
 } from 'lucide-react'
 import { useEffect, useRef, useState } from 'react'
 
@@ -41,7 +37,7 @@ export default function Page({
 
   /* 初始化时间 */
   useEffect(() => {
-    setInitialTime(Date.now() - 10 * 24 * 60 * 60 * 1000)
+    setInitialTime(Date.now() - 2 * 24 * 60 * 60 * 1000)
   }, [])
 
   /* 加载摄像头信息 */
@@ -83,7 +79,7 @@ export default function Page({
   }
   // 初始化完成
   return (
-    <div className="mx-auto grid max-w-6xl gap-4 rounded-xl md:p-8">
+    <div className="max-w-8xl mx-auto grid w-full gap-4 rounded-xl md:w-2/3 md:p-8">
       <h1 className="mb-6 text-3xl font-bold">{cameraInfo?.Name || ''}</h1>
       {/*播放器*/}
       <CameraPlayer