+++ web/src/components/main/nav-cameras.tsx
@@ -1,12 +1,6 @@
 'use client'
 
-import {
-  Camera as CameraIcon,
-  Gauge,
-  MonitorCog,
-  MonitorPlay,
-  MoreHorizontal,
-} from 'lucide-react'
+import { Camera as CameraIcon, Clapperboard, MonitorCog, MonitorPlay, MoreHorizontal } from 'lucide-react'
 
 import { components } from '@/api/schema'
 import {
@@ -14,7 +8,7 @@ import {
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
-  DropdownMenuTrigger,
+  DropdownMenuTrigger
 } from '@/components/ui/dropdown-menu'
 import {
   SidebarGroup,
@@ -23,7 +17,7 @@ import {
   SidebarMenuAction,
   SidebarMenuButton,
   SidebarMenuItem,
-  useSidebar,
+  useSidebar
 } from '@/components/ui/sidebar'
 import { useLocale } from 'next-intl'
 import Link from 'next/link'
@@ -66,10 +60,10 @@ export function NavCameras({ cameras }: { cameras: Camera[] }) {
                 side={isMobile ? 'bottom' : 'right'}
                 align={isMobile ? 'end' : 'start'}
               >
-                <Link href={`/${locale}/${item.Id}/dashboard`}>
+                <Link href={`/${locale}/${item.Id}/segments`}>
                   <DropdownMenuItem>
-                    <Gauge className="text-muted-foreground" />
-                    <span>Dashboard</span>
+                    <Clapperboard className="text-muted-foreground" />
+                    <span>Segments</span>
                   </DropdownMenuItem>
                 </Link>
                 <Link href={`/${locale}/${item.Id}/playback`}>
@@ -79,10 +73,10 @@ export function NavCameras({ cameras }: { cameras: Camera[] }) {
                   </DropdownMenuItem>
                 </Link>
                 <DropdownMenuSeparator />
-                <Link href={`/${locale}/${item.Id}/manage`}>
+                <Link href={`/${locale}/${item.Id}/settings`}>
                   <DropdownMenuItem>
                     <MonitorCog className="text-muted-foreground" />
-                    <span>Manage</span>
+                    <span>Settings</span>
                   </DropdownMenuItem>
                 </Link>
               </DropdownMenuContent>